requirements.txt
SQLAlchemy>=1.4.45
pandas>=1.5.0
tabulate>=0.9.0
pyspark>=3.3.1
peewee>=3.15.3
numpy>=1.23.4
matplotlib>=3.6.1
ipython>=8.5.0
python-dotenv>=0.21.0
psycopg2-binary>=2.9.5
